% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add and Sub.R
\name{jmath}
\alias{jmath}
\alias{jadd}
\alias{jsub}
\title{Justin's function}
\usage{
jadd(x, y)

jsub(x, y)
}
\value{
\code{x + y}

\code{x - y}
}
\description{
Simple arithmetic functions

Add \code{x} and \code{y}

Subtract \code{y} from \code{x}
}
\examples{
jadd(1,2)
jsub(1,2)
}
